pyproject.toml
[tool.poetry]
name = "twitter"
version = "0.1.0"
description = ""
authors = ["Chi Wai CHAN <ray.chan@hennge.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
google-api-python-client = "^2.11.0"
google-auth = "^1.32.1"
google-auth-oauthlib = "^0.4.4"
google-auth-httplib2 = "^0.1.0"
redis = "^3.5.3"
boto3 = "^1.17.108"
Werkzeug = "^2.0.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
twitter = "twitter.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
